move the tiptap model to the BibleWell project, and rename the file
