+++ src/BibleWell.App/Views/Components/TiptapRendererView.axaml.cs
@@ -5,7 +5,6 @@ using Avalonia.Controls.Documents;
 using Avalonia.Layout;
 using Avalonia.Media;
 using BibleWell.App.ViewModels.Components;
-using BibleWell.Models;
 
 namespace BibleWell.App.Views.Components;
