+++ src/BibleWell/BibleWell.csproj
@@ -8,4 +8,8 @@
         <ProjectReference Include="..\BibleWell.Utility\BibleWell.Utility.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Models\" />
+    </ItemGroup>
+
 </Project>
